pyproject.toml
[project]
name = "lovli"
version = "0.1.0"
description = "Legal RAG assistant for Norwegian laws using LangChain, Qdrant, and GLM-4.7"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "Andreas Ramsli", email = "andreas@example.com"}
]
keywords = ["legal", "rag", "norwegian", "langchain", "qdrant"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "langchain>=0.1.0",
    "langchain-core>=0.1.0",
    "langchain-community>=0.0.20",
    "langchain-classic>=1.0.0",
    "langchain-openai>=0.1.0",
    "langchain-qdrant>=1.0.0",
    "qdrant-client>=1.7.0",
    "sentence-transformers>=2.3.0",
    "beautifulsoup4>=4.12.0",
    "lxml>=5.1.0",
    "streamlit>=1.29.0",
    "python-dotenv>=1.0.0",
    "pydantic-settings>=2.1.0",
    "pydantic>=2.5.0",
    "langsmith>=0.3.11",
    "openevals",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "black>=23.12.0",
    "ruff>=0.1.0",
]
test = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "scipy>=1.11.0",  # For sparse vector format tests
]

[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 100
target-version = ['py311']

[tool.ruff]
line-length = 100
target-version = "py311"
